Fix hrtime(true) to return float per php-src (#12779)
hrtime(as_number: true) now returns double like microtime(true), matching
ext/standard/hrtime.c. VM, JIT lowering, and HrtimeJitHelper::nsFloat share
the PHP SSOT path.

+++ ext/standard/VmHrtime.php
@@ -15,13 +15,13 @@ final class VmHrtime
     private const NS_PER_SEC = 1_000_000_000;
 
     /**
-     * @return int|array{0: int, 1: int}
+     * @return float|array{0: int, 1: int}
      */
     public static function hrtime(bool $asNumber = false)
     {
         [$sec, $nsec] = VmHrtimeNative::readMonotonic();
         if ($asNumber) {
-            return (int) ($sec * self::NS_PER_SEC + $nsec);
+            return (float) ($sec * self::NS_PER_SEC + $nsec);
         }
 
         return [$sec, $nsec];
